fixed textarea test
